requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure.cosmos
azure-storage-blob
tensorflow
tensorflow-hub==0.12.0
pillow